+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Updated build process.
 
+## Added
+
+- Added support for files without comment symbols like `.txt` files.
+
 ## [1.0.4] - 2024-11-29
 
 ### Changed